Fix #74
